requirements.txt
# PyTorch and Torchvision installed from source for Jetson-Nano
# Please comment torch & torchvision when working on Jetson-Nano

# torch==1.6.0
https://download.pytorch.org/whl/cpu/torch-1.5.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
# torchvision==0.7.0
https://download.pytorch.org/whl/cpu/torchvision-0.6.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
flask
numpy
Pillow
Click>=7.1.2
itsdangerous==1.1.0
Jinja2>=2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
gunicorn>=20.0.4